corretto parsefloat su distance

+++ js/script.js
@@ -8,8 +8,8 @@ const inputDiscount = document.querySelector('main div button')
 inputDiscount.addEventListener('click', function(){
 
     /* inserite variabili dentro ricaricare sempre il loro valore all'input sul button rilevato dal listener */
-    let userAge = parseFloat(document.getElementById('user-Age').value)
-    let userDistance = parseInt(document.getElementById('user-distance').value)
+    let userAge = parseInt(document.getElementById('user-Age').value)
+    let userDistance = parseFloat(document.getElementById('user-distance').value)
     /* mostrato il valore in console */
 
     console.log(`User age: ${userAge}`)
